🚚 move: createDOM 코드 분리
