+++ src/main.tsx
@@ -1,6 +1,5 @@
-import { createRoot } from './react';
+import { render } from './react';
 import { App } from './App';
 import './style.css';
 
-const root = createRoot(document.querySelector('#root')!);
-root.render(App);
+render(App, document.querySelector('#root')!);
